Rename and refactor SocialPost component
